move legal container to atoms
