+++ frontend/src/components/atoms/LegalContainer.vue
